// ET-Managed/src/app/services/stores/models/system-data-actions.model.ts
export enum SystemDataActionType {
    GET_SYSTEM_DATA = '[SYSTEM_DATA] GET_SYSTEM_DATA ',
    GET_SYSTEM_DATA_SUCCESS = '[SYSTEM_DATA] GET_SYSTEM_DATA_SUCCESS',
    GET_SYSTEM_DATA_FAILURE = '[SYSTEM_DATA] GET_SYSTEM_DATA_FAILURE',
}

export enum NotificationActionType {
    HIDE_NOTIFICATIONS = '[NOTIFICATION] HIDE_NOTIFICATIONS',
    DELETE_NOTIFICATIONS = '[NOTIFICATION] DELETE_NOTIFICATIONS',
    DELETE_NOTIFICATION_BY_ID = '[NOTIFICATION] DELETE_NOTIFICATION_BY_ID',
    UPDATE_NOTIFICATION_IS_NEW = '[NOTIFICATION] UPDATE_NOTIFICATION_IS_NEW',
    UPDATE_ALL_NOTIFICATIONS_ARE_INACTIVE_AND_NOT_EXPANDED = '[NOTIFICATION] UPDATE_ALL_NOTIFICATIONS_ARE_INACTIVE_AND_NOT_EXPANDED',
    UPDATE_NOTIFICATION_ITEM_IS_ACTIVE = '[NOTIFICATION] UPDATE_NOTIFICATION_ITEM_IS_ACTIVE',
    UPDATE_NOTIFICATION_ITEM_IS_EXPANDED = '[NOTIFICATION] UPDATE_NOTIFICATION_ITEM_IS_EXPANDED'
}